+++ lib/jazzy/source_module.rb
@@ -4,30 +4,27 @@ require 'uri'
 
 require 'jazzy/config'
 require 'jazzy/source_declaration'
+require 'jazzy/source_host'
 
 module Jazzy
   class SourceModule
     attr_accessor :name
-    attr_accessor :root_path
     attr_accessor :docs
     attr_accessor :doc_coverage
     attr_accessor :doc_structure
     attr_accessor :author_name
-    attr_accessor :github_url
-    attr_accessor :github_file_prefix
     attr_accessor :author_url
     attr_accessor :dash_url
+    attr_accessor :host
 
     def initialize(options, docs, doc_structure, doc_coverage)
       self.docs = docs
-      self.root_path = options.source_directory
       self.doc_structure = doc_structure
       self.doc_coverage = doc_coverage
       self.name = options.module_name
       self.author_name = options.author_name
-      self.github_url = options.github_url
-      self.github_file_prefix = options.github_file_prefix
       self.author_url = options.author_url
+      self.host = SourceHost.create(options)
       return unless options.dash_url
 
       self.dash_url =